lua/utils: `network` module for URI utilities
This patch adds a network module, which provides utility functions
for validating and handling network-related objects, such as URIs.

The `uri_is_suitable_to_connect` function validates a URI for use
in creating a client socket with `connect()`. It ensures the URI
does not use INADDR_ANY (`0.0.0.0`), in6addr_any (`::`), or a zero
TCP port. Returns `true` if valid or `false, err` with an error
message otherwise.

Needed for #10691

NO_DOC=internal
NO_CHANGELOG=internal
NO_TEST=internal
