+++ src/box/lua/config/instance_config.lua
@@ -7,6 +7,7 @@ local fio = require('fio')
 local file = require('internal.config.utils.file')
 local log = require('internal.config.utils.log')
 local funcutils = require('internal.config.utils.funcutils')
+local network = require('internal.config.utils.network')
 
 -- List of annotations:
 --
@@ -128,47 +129,6 @@ local function validate_uuid_str(data, w)
     end
 end
 
--- Accepts an uri object (one produced by urilib.parse()).
---
--- Performs several checks regarding ability to use the URI to
--- create a client socket. IOW, to call connect() on it.
---
--- The function returns `true` if the URI is OK to connect and
--- `false, err` otherwise.
---
--- If the URI doesn't fit the given criteria an error is raised.
--- The function returns nothing otherwise.
---
--- The following checks are performed:
---
--- * INADDR_ANY IPv4 address (0.0.0.0) or in6addr_any IPv6 address
---   (::) in the host part of the URI.
---
---   It means 'bind to all interfaces' for the bind() call, but it
---   has no meaning at the connect() call on a client.
--- * Zero TCP port (service part of the URI).
---
---   It means 'bind to a random free port' for the bind() call,
---   but it has no meaning at the connect() call on a client.
-local function uri_is_suitable_to_connect(uri)
-    assert(uri ~= nil)
-
-    if uri.ipv4 == '0.0.0.0' then
-        return false, 'INADDR_ANY (0.0.0.0) cannot be used to create ' ..
-            'a client socket'
-    end
-    if uri.ipv6 == '::' then
-        return false, 'in6addr_any (::) cannot be used to create a client ' ..
-            'socket'
-    end
-    if uri.service == '0' then
-        return false, 'An URI with zero port cannot be used to create ' ..
-            'a client socket'
-    end
-
-    return true
-end
-
 -- Verify 'uri' field in iproto.listen and iproto.advertise.* options.
 local function validate_uri_field(has_login_field, used_to_connect)
     return function(data, w)
@@ -194,7 +154,7 @@ local function validate_uri_field(has_login_field, used_to_connect)
                 "field, not as the part of URI")
         end
         if used_to_connect then
-            local ok, err = uri_is_suitable_to_connect(uri)
+            local ok, err = network.uri_is_suitable_to_connect(uri)
             if not ok then
                 w.error("bad URI %q: %s", data, err)
             end
@@ -216,7 +176,7 @@ local function find_suitable_uri_to_connect(listen, opts)
         assert(u.uri ~= nil)
         local uri = urilib.parse(u.uri)
         if uri ~= nil then
-            local ok, err = uri_is_suitable_to_connect(uri)
+            local ok, err = network.uri_is_suitable_to_connect(uri)
             if ok then
                 -- The urilib.format() call has the second optional
                 -- argument `write_password`. Let's assume that the
